Banner GFX Format
This commit introduces the BannerGfxFormat class and the required cast operators to turn a byte into an instance of this class.

By using this representation of the data, we have a better way of actually reading the various flags stored inside that byte and represent them properly inside the source code.